Fix the whole thing.
Well, for starters, this code doesn't even work, let alone do what you want it to do.  This isn't even half way finished.  The variables are not going to work the way that they are written at all and there is no way that you can declare public and private the way that you have here.  This isn't going to work and needs to be thrown out and start over.